# src/main/resources/com/sonicle/webtop/contacts/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Contacts",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is Contacts service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    }
  ],
  "paths": {
...
    "/me/categories": {
      "get": {
        "summary": "List Categories",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResult"
                }
              }
            }
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "operationId": "listCategories",
        "description": "Returns a list of categories readable by the current user: this includes both personal and incoming shared categories.",
        "parameters": [
          {
            "$ref": "#/components/parameters/dollar_filter"
          },
          {
            "$ref": "#/components/parameters/dollar_select"
          },
          {
            "$ref": "#/components/parameters/dollar_orderBy"
          },
          {
            "$ref": "#/components/parameters/dollar_pageNo"
          },
          {
            "$ref": "#/components/parameters/dollar_pageSize"
          },
          {
            "$ref": "#/components/parameters/dollar_returnCount"
          }
        ]
      },
      "post": {
        "summary": "Add a Category",
        "operationId": "addCategory",
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "user_id",
            "description": "The ID of a user"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryBase"
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Adds new category specifying the owning user ID. If no user ID is provided, the owner will be the current user."
      }
    },
    "/me/categories/{category_id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "category_id",
          "in": "path",
          "required": true,
          "description": "The ID of a category"
        }
      ],
      "get": {
        "summary": "Get a category",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Category not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "operationId": "getCategory",
        "description": "Gets the specified category given its ID."
      },
      "put": {
        "summary": "Update a category",
        "operationId": "updateCategory",
        "responses": {
          "200": {
            "description": "Category updated"
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Category not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "tags": [
          "me"
        ],
        "x-codegen-request-body-name": "body",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryBase"
              }
            }
          }
        },
        "description": "Update the specified category given its ID. You can choose to update the entire object or only a subset of data."
      },
      "delete": {
        "summary": "Delete a category",
        "operationId": "deleteCategory",
        "responses": {
          "204": {
            "description": "Category deleted"
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Category not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "tags": [
          "me"
        ],
        "description": "Delete a category given its ID."
      }
    },
    "/me/categories/{category_id}/contacts": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "category_id",
          "in": "path",
          "required": true,
          "description": "The ID of a category"
        }
      ],
      "get": {
        "summary": "List contacts",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsResult"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "operationId": "listCategoryContacts",
        "parameters": [
          {
            "$ref": "#/components/parameters/dollar_filter"
          },
          {
            "$ref": "#/components/parameters/dollar_select"
          },
          {
            "$ref": "#/components/parameters/dollar_orderBy"
          },
          {
            "$ref": "#/components/parameters/dollar_pageNo"
          },
          {
            "$ref": "#/components/parameters/dollar_pageSize"
          },
          {
            "$ref": "#/components/parameters/dollar_returnCount"
          }
        ],
        "description": "Returns a list of contacts from specified category."
      },
      "post": {
        "summary": "Add a contact",
        "tags": [
          "me"
        ],
        "operationId": "addCategoryContact",
        "responses": {
          "201": {
            "description": "Contact created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "404": {
            "description": "Category not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactEx"
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Adds new contact into given category."
      }
    },
    "/me/categories/{category_id}/contacts/delta": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "category_id",
          "in": "path",
          "required": true,
          "description": "The ID of a category"
        }
      ],
      "get": {
        "summary": "List changes on contacts collection",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsResultDelta"
                }
              }
            }
          }
        },
        "operationId": "listCategoryContactsDelta",
        "parameters": [
          {
            "$ref": "#/components/parameters/dollar_syncToken"
          },
          {
            "$ref": "#/components/parameters/dollar_select"
          }
        ],
        "description": "Get a set of contacts that have been added, deleted, or updated in a specified category, starting from a precise instant identified by a given syncToken."
      }
    },
    "/me/contacts": {
      "post": {
        "summary": "Add a contact",
        "operationId": "addContact",
        "responses": {
          "201": {
            "description": "Contact created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        },
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "category_id",
            "required": true,
            "description": "The category ID where the contact will be added into"
          }
        ],
        "x-codegen-request-body-name": "body",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactEx"
              }
            }
          }
        },
        "description": "Adds new contact into specified category."
      }
    },
    "/me/contacts/{contact_id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "contact_id",
          "in": "path",
          "required": true,
          "description": "The ID of a contact"
        }
      ],
      "get": {
        "summary": "Get a contact",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "operationId": "getContact",
        "parameters": [
          {
            "$ref": "#/components/parameters/dollar_select"
          }
        ],
        "description": "Gets the specified contact given its ID."
      },
      "put": {
        "summary": "Update a contact",
        "operationId": "updateContact",
        "responses": {
          "200": {
            "description": "Contact updated",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactEx"
              }
            }
          }
        },
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/dollar_update"
          }
        ],
        "x-codegen-request-body-name": "body",
        "description": "Update the specified contact given its ID. You can choose to update the entire object or only a subset of data."
      },
      "delete": {
        "summary": "Delete a contact",
        "operationId": "deleteContact",
        "responses": {
          "204": {
            "description": "Contact deleted"
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "tags": [
          "me"
        ],
        "description": "Delete a contact given its ID."
      }
    },
    "/me/contacts/{contact_id}/picture": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "contact_id",
          "in": "path",
          "required": true,
          "description": "The ID of a contact"
        }
      ],
      "get": {
        "summary": "Get a contact picture metadata",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPictureMeta"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {}
          }
        },
        "operationId": "getContactPictureMeta",
        "description": "Gets the contact's picture metatada of a given contact ID."
      },
      "put": {
        "summary": "Set contact picture",
        "tags": [
          "me"
        ],
        "operationId": "setContactPicture",
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "description": "Assign a picture to the given contact ID. The picture should be in binary. It replaces any existing photo for that contact.",
        "requestBody": {
          "content": {
            "image/*": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "description": "The binary data of the picture."
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "Content-Type",
            "description": "The picture media-type."
          }
        ]
      },
      "delete": {
        "summary": "",
        "operationId": "clearContactPicture",
        "responses": {
          "204": {
            "description": "Contact's picture deleted"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "description": "Deletes the contact's picture of a given contact ID.",
        "tags": [
          "me"
        ]
      }
    },
    "/me/contacts/{contact_id}/picture/$value": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "contact_id",
          "in": "path",
          "required": true,
          "description": "The ID of a contact"
        }
      ],
      "get": {
        "summary": "Get contact picture binary data",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "operationId": "getContactPictureBytes",
        "description": "Gets the contact's picture raw data (binary), if any, for a given contact ID."
      }
    }
  },
  "components": {
    "schemas": {
...
      "CategoriesResult": {
        "title": "CategoriesResult",
        "x-stoplight": {
          "id": "9cwwa9calzrpg"
        },
        "type": "object",
        "description": "Represents a response object for a collection of categories.",
        "x-class-extra-annotation": "@JsonInclude(JsonInclude.Include.NON_NULL)",
        "properties": {
          "totalCount": {
            "type": "integer",
            "x-stoplight": {
              "id": "y3ogz86j919ct"
            },
            "description": "Items total count, usually used for pagination.",
            "readOnly": true
          },
          "items": {
            "type": "array",
            "x-stoplight": {
              "id": "2sgsj1752mcd3"
            },
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
...
      "dollar_select": {
        "name": "$select",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "example": "name, description"
        },
        "description": "List (comma-separated) of field names to include in resulting items. Optional."
      },
      "dollar_orderBy": {
        "name": "$order_by",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "example": "name ASC, year DESC"
        },
        "description": "List (comma-separated) of field names and direction (ASC or DESC) to sort resulting items. Optional."
      },
      "dollar_pageNo": {
        "name": "$page_no",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "description": "The page number to return, providing a value actually activates pagination. Optional."
      },
      "dollar_pageSize": {
        "name": "$page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "description": "How many items to return when paginating. Defaults to 50."
      },
      "dollar_returnCount": {
        "name": "$return_count",
        "in": "query",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "Specifies whether to compute and return the full count of a list of items. Useful when dealing with paginated data. Optional."
      },
      "dollar_filter": {
        "name": "$filter",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A RSQL filter query to filter out resulting items. Optional."
      },
...
    },
...
  },
...
}